Finished Project Webpage Lily Chau
The website is not completley finished, but meets all the requirements. I understand the use of html and css and am capable of finishing the website in the future. The completed pages are index.html, auggies.html, and starbucks.html (as well as style.css, Aug.css, and star.css which are used to design). My outline was submitted through Google Classroom.